# Detailed/03-OpenAPI/med-rms-openapi.yaml
...
openapi: 3.0.3
info:
  title: Med-RMS 医疗器械需求管理系统 API
  description: 'Med-RMS (Medical Device Requirement Management System) RESTful API 规范。

    涵盖需求管理、追溯管理、变更管理、合规管理、电子签名、风险管理、项目管理、报表仪表盘及系统管理模块。

    '
  version: 1.1.0
  contact:
    name: Med-RMS 开发团队
  license:
    name: Proprietary
servers:
- url: http://localhost:8080/api/v1
  description: 开发环境
...
security:
- BearerAuth: []
paths:
...
  /changes:
    post:
      operationId: createChangeRequest
      tags:
      - 变更管理
      summary: 创建变更请求
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestCreateDTO'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ChangeRequestVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: chg:create
    get:
      operationId: getChangeRequestList
      tags:
      - 变更管理
      summary: 查询变更列表
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/SizeParam'
      - $ref: '#/components/parameters/SortParam'
      - name: projectId
        in: query
        description: 项目ID
        schema:
          type: string
      - name: status
        in: query
        description: 变更状态
        schema:
          $ref: '#/components/schemas/ChangeStatus'
      - name: changeType
        in: query
        description: 变更类型
        schema:
          $ref: '#/components/schemas/ChangeType'
      - name: keyword
        in: query
        description: 搜索关键词
        schema:
          type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PageChangeRequestVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: chg:list
...
  /dashboard:
    get:
      operationId: getDashboard
      tags:
      - 报表仪表盘
      summary: 获取仪表盘数据
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/DashboardVO'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: report:dashboard
  /dashboard/layout:
    put:
      operationId: saveDashboardLayout
      tags:
      - 报表仪表盘
      summary: 保存仪表盘布局
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutDTO'
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-required-permission: report:dashboard
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: 每页大小（默认20，最大100）
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: 排序（field,direction 如 createdAt,desc）
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证（Token缺失或过期）
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 系统内部错误
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: success
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
    PageMeta:
      type: object
      required:
      - page
      - size
      - total
      - totalPages
      properties:
        page:
          type: integer
          description: 当前页码
          example: 1
        size:
          type: integer
          description: 每页大小
          example: 20
        total:
          type: integer
          format: int64
          description: 总记录数
          example: 156
        totalPages:
          type: integer
          description: 总页数
          example: 8
    ErrorResponse:
      type: object
      required:
      - errorCode
      - errorMessage
      properties:
        errorCode:
          type: string
          description: 错误码
          example: SY0100
        errorMessage:
          type: string
          description: 错误消息
          example: 请求参数校验失败
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: 字段级错误详情
...
    ChangeStatus:
      type: string
      enum:
      - Draft
      - Submitted
      - UnderAnalysis
      - PendingApproval
      - Approved
      - Rejected
      - Executing
      - Verifying
      - Completed
      - Closed
      description: 变更状态
    ChangeType:
      type: string
      enum:
      - Correction
      - Enhancement
      - Adaptation
      - Prevention
      description: 变更类型
...
    ChangeRequestCreateDTO:
      type: object
      required:
      - projectId
      - title
      - changeType
      - reason
      properties:
        projectId:
          type: string
          format: int64
          description: 项目ID
        title:
          type: string
          description: 变更标题
          maxLength: 200
        changeType:
          $ref: '#/components/schemas/ChangeType'
        reason:
          type: string
          description: 变更原因
          maxLength: 1000
        affectedItems:
          type: array
          items:
            type: object
            properties:
              itemType:
                type: string
                description: 影响项类型
              itemId:
                type: string
                format: int64
                description: 影响项ID
          description: 受影响项列表
        proposedSolution:
          type: string
          description: 提议方案
    ChangeRequestVO:
      type: object
      properties:
        id:
          type: string
          format: int64
        changeCode:
          type: string
          description: 变更编号
        projectId:
          type: string
          format: int64
        projectName:
          type: string
        title:
          type: string
        changeType:
          $ref: '#/components/schemas/ChangeType'
        status:
          $ref: '#/components/schemas/ChangeStatus'
        reason:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageChangeRequestVO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestVO'
        pagination:
          $ref: '#/components/schemas/PageMeta'
...
    DashboardVO:
      type: object
      properties:
        widgets:
          type: array
          items:
            type: object
            properties:
              widgetId:
                type: string
              type:
                type: string
              title:
                type: string
              position:
                type: object
                properties:
                  x:
                    type: integer
                  y:
                    type: integer
                  width:
                    type: integer
                  height:
                    type: integer
              data:
                type: object
                additionalProperties: true
        summary:
          type: object
          properties:
            totalProjects:
              type: integer
            activeRequirements:
              type: integer
            pendingChanges:
              type: integer
            openRisks:
              type: integer
            complianceRate:
              type: number
              format: double
    LayoutDTO:
      type: object
      required:
      - widgets
      properties:
        widgets:
          type: array
          items:
            type: object
            required:
            - widgetId
            - position
            properties:
              widgetId:
                type: string
              position:
                type: object
                required:
                - x
                - y
                - width
                - height
                properties:
                  x:
                    type: integer
                  y:
                    type: integer
                  width:
                    type: integer
                  height:
                    type: integer
          description: 仪表盘组件布局
...